moved shadow_btn and flat_btn to won widgets
