+++ lib/screens/success_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:namer_app/widgets/category_pill.dart';
 import 'package:namer_app/widgets/feedback_row.dart';
 import 'package:namer_app/widgets/reflection.dart';
-import 'package:namer_app/buttons/save_btn.dart';
+import 'package:namer_app/buttons/shadow_btn.dart';
 import 'package:namer_app/services/database_service.dart';
 
 class SuccessScreen extends StatefulWidget {
@@ -124,7 +124,7 @@ class _SuccessScreenState extends State<SuccessScreen> {
             ),
             Padding(
               padding: const EdgeInsets.all(16.0),
-              child: SaveBtn(onPressed: _onSavePressed),
+              child: ShadowBtn(onPressed: _onSavePressed, btnText: "Save"),
             ),
           ],
         ),
